+++ NEWS.md
@@ -1,7 +1,8 @@
 # MSnbase 2.3
 
 ## Changes in version 2.3.3
-- Nothing yet
+- Rewrite `getColsFromPattern` and `getRowsFromPattern` and add unit tests
+    <2017-05-11 Thu>.
 
 ## Changes in version 2.3.2
 - Rewrite `utils.clean`. It now keeps just the zeros in the direct